requirements.txt
flask
flasgger
flask_sqlalchemy
pymysql
gunicorn
line-bot-sdk
python-dotenv

